refactor: mvvm refactor
